# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WikiKit API",
    "version": "0.22.0",
    "description": "Headless, AI-native knowledge system for humans and agents. Markdown-first knowledge in; structured, cited, review-gated knowledge out. Scopes: knowledge:read | knowledge:propose | knowledge:review | knowledge:approve | admin (approve implies review). MCP (Streamable HTTP) is available at POST /mcp, outside this REST surface; remote MCP clients use OAuth 2.1 discovery, PKCE and explicitly requested scopes."
  },
  "servers": [{ "url": "/" }],
  "paths": {
...
    "/v1/spaces/{space}/concepts/{slug}": {
      "get": {
        "summary": "Read a concept: markdown + claims + citations + relations",
        "operationId": "getConcept",
        "responses": {
          "200": {
            "description": "Concept",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/zConceptResponse" } } }
          },
          "400": {
            "description": "bad_request — request failed schema validation",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/zErrorEnvelope" } } }
          },
          "401": {
            "description": "unauthorized — missing, unknown or revoked API key",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/zErrorEnvelope" } } }
          },
          "403": {
            "description": "insufficient_scope — key lacks the required scope or is scoped to another space",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/zErrorEnvelope" } } }
          },
          "404": {
            "description": "not_found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/zErrorEnvelope" } } }
          },
          "500": {
            "description": "internal_error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/zErrorEnvelope" } } }
          }
        },
        "parameters": [
          {
            "name": "space",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "pattern": "^[a-z0-9][a-z0-9-]{0,62}$" }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "pattern": "^[a-z0-9][a-z0-9-]{0,126}$" }
          }
        ],
        "security": [{ "oauth2": [] }, { "bearerAuth": [] }, { "apiKey": [] }],
        "x-required-scope": "knowledge:read"
      }
    },
...
    "/v1/spaces/{space}/stats/http": {
      "get": {
        "summary": "Privacy-safe HTTP requests, outcomes, exact-window adoption and latency",
        "operationId": "httpUsageStats",
        "responses": {
          "200": {
            "description": "HTTP usage statistics",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/zUsageStatsResponse" } } }
          },
          "400": {
            "description": "Invalid usage statistics query",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/zErrorEnvelope" } } }
          },
          "401": {
            "description": "unauthorized — missing, unknown or revoked API key",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/zErrorEnvelope" } } }
          },
          "403": {
            "description": "insufficient_scope — key lacks the required scope or is scoped to another space",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/zErrorEnvelope" } } }
          },
          "404": {
            "description": "not_found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/zErrorEnvelope" } } }
          },
          "500": {
            "description": "internal_error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/zErrorEnvelope" } } }
          }
        },
        "parameters": [
          {
            "name": "space",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "pattern": "^[a-z0-9][a-z0-9-]{0,62}$" }
          },
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["hour", "day", "month", "year"] }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
            }
          },
          { "name": "tz", "in": "query", "required": false, "schema": { "type": "string", "const": "UTC" } },
          {
            "name": "traffic_class",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["organic", "synthetic", "internal", "all"] }
          },
          { "name": "group_by", "in": "query", "required": false, "schema": { "type": "string", "maxLength": 200 } }
        ],
        "security": [{ "oauth2": [] }, { "bearerAuth": [] }, { "apiKey": [] }],
        "x-required-scope": "knowledge:read"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "OAuth 2.1 authorization code with PKCE-S256 for interactive clients.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/v1/oauth/authorize",
            "tokenUrl": "/v1/oauth/token",
            "scopes": {
              "knowledge:read": "Read spaces, sources, concepts, search and reviewed proposals",
              "knowledge:propose": "Create ingest jobs and staged proposals",
              "knowledge:review": "Inspect and start human proposal review",
              "knowledge:approve": "Approve or reject proposals after explicit human review"
            }
          }
        }
      },
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "Authorization: Bearer wk_..." },
      "apiKey": { "type": "apiKey", "in": "header", "name": "X-API-Key" }
    },
    "schemas": {
...
      "zErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" },
          "request_id": { "type": "string" },
          "next_best_actions": { "type": "array", "items": { "type": "string" } }
        },
        "required": ["error", "code", "request_id"],
        "additionalProperties": {}
      },
...
      "zConceptResponse": {
        "type": "object",
        "properties": {
          "slug": { "type": "string" },
          "title": { "type": "string" },
          "summary": { "type": "string" },
          "markdown": { "type": "string" },
          "rev": { "type": "integer", "minimum": -9007199254740991, "maximum": 9007199254740991 },
          "updated_at": { "type": "string" },
          "claims": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "subject": { "type": "string" },
                "predicate": { "type": "string" },
                "object": { "type": "string" },
                "status": { "type": "string", "enum": ["verified", "disputed", "deprecated"] },
                "confidence": { "type": "number" },
                "citations": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "source_id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "quote": { "type": "string" },
                      "locator": { "type": "string" }
                    },
                    "required": ["source_id", "quote", "locator"],
                    "additionalProperties": false
                  }
                }
              },
              "required": ["id", "subject", "predicate", "object", "status", "confidence", "citations"],
              "additionalProperties": false
            }
          },
          "relations": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "to_slug": { "type": "string" },
                "kind": { "type": "string", "enum": ["related", "part_of", "depends_on", "contradicts", "supersedes"] },
                "space": { "anyOf": [{ "type": "string" }, { "type": "null" }] }
              },
              "required": ["to_slug", "kind", "space"],
              "additionalProperties": false
            }
          },
          "agent_meta": { "type": "object", "propertyNames": { "type": "string" }, "additionalProperties": {} }
        },
        "required": ["slug", "title", "summary", "markdown", "rev", "updated_at", "claims", "relations", "agent_meta"],
        "additionalProperties": false
      },
...
    }
  }
}